Revert "DTSCCI-958 Refactor MediationJsonService (#5624)"
This reverts commit fadd6a35362ea93bcc96fd86088b8a35f1578014.
