+++ src/main/java/uk/gov/hmcts/reform/civil/service/mediation/MediationJsonService.java
@@ -2,48 +2,62 @@ package uk.gov.hmcts.reform.civil.service.mediation;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.jetbrains.annotations.Nullable;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.ccd.model.OrganisationPolicy;
 import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.model.Party;
+import uk.gov.hmcts.reform.civil.model.UnavailableDate;
 import uk.gov.hmcts.reform.civil.model.caseflags.FlagDetail;
 import uk.gov.hmcts.reform.civil.model.caseflags.PartyFlags;
 import uk.gov.hmcts.reform.civil.model.citizenui.MediationLiPCarm;
+import uk.gov.hmcts.reform.civil.model.common.Element;
+import uk.gov.hmcts.reform.civil.model.dq.Applicant1DQ;
+import uk.gov.hmcts.reform.civil.model.dq.Applicant2DQ;
 import uk.gov.hmcts.reform.civil.model.dq.HearingSupport;
+import uk.gov.hmcts.reform.civil.model.dq.Respondent1DQ;
+import uk.gov.hmcts.reform.civil.model.dq.Respondent2DQ;
 import uk.gov.hmcts.reform.civil.model.dq.VulnerabilityQuestions;
 import uk.gov.hmcts.reform.civil.model.dq.WelshLanguageRequirements;
 import uk.gov.hmcts.reform.civil.model.mediation.MediationAvailability;
 import uk.gov.hmcts.reform.civil.model.mediation.MediationContactInformation;
-import uk.gov.hmcts.reform.civil.service.mediation.helpers.PartyDetailsPopulator;
-import uk.gov.hmcts.reform.civil.service.mediation.helpers.RepresentedLitigantPopulator;
-import uk.gov.hmcts.reform.civil.service.mediation.helpers.UnrepresentedLitigantPopulator;
+import uk.gov.hmcts.reform.civil.prd.model.Organisation;
+import uk.gov.hmcts.reform.civil.service.OrganisationService;
 
+import java.time.LocalDate;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Stream;
+import java.util.Locale;
+import java.util.Optional;
 
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.NO;
 import static uk.gov.hmcts.reform.civil.enums.YesOrNo.YES;
 import static uk.gov.hmcts.reform.civil.enums.dq.Language.BOTH;
 import static uk.gov.hmcts.reform.civil.enums.dq.Language.WELSH;
+import static uk.gov.hmcts.reform.civil.enums.dq.UnavailableDateType.DATE_RANGE;
+import static uk.gov.hmcts.reform.civil.enums.dq.UnavailableDateType.SINGLE_DATE;
 import static uk.gov.hmcts.reform.civil.utils.CaseFlagsHearingsUtils.getAllActiveCaseLevelFlags;
 import static uk.gov.hmcts.reform.civil.utils.CaseFlagsHearingsUtils.getAllActiveFlags;
+import static uk.gov.hmcts.reform.civil.utils.ElementUtils.unwrapElements;
 
 @Slf4j
 @Service
 @AllArgsConstructor
 public class MediationJsonService {
 
-    private final PartyDetailsPopulator partyDetailsPopulator;
-    private final RepresentedLitigantPopulator representedLitigantPopulator;
-    private final UnrepresentedLitigantPopulator unrepresentedLitigantPopulator;
+    private final OrganisationService organisationService;
+
+    private static final String PAPER_RESPONSE = "N";
+    private static final String DATE_FORMAT = "yyyy-MM-dd";
 
     public MediationCase generateJsonContent(CaseData caseData) {
-        log.info("Generate JSON content for case ID {}", caseData.getCcdCaseReference());
         List<MediationLitigant> litigantList = new ArrayList<>();
 
+        // caseFlags
         boolean activeCaseFlags = buildCaseFlags(caseData);
 
+        // litigants
         buildLitigantFields(caseData, litigantList);
 
         return MediationCase.builder()
@@ -57,78 +71,62 @@ public class MediationJsonService {
     }
 
     private boolean buildCaseFlags(CaseData caseData) {
-        log.info("Build case flags for case ID {}", caseData.getCcdCaseReference());
-        return Stream.of(
-            checkForActiveCaseFlags(caseData),
-            checkApplicant1DQRequirements(caseData),
-            checkRespondent1DQRequirements(caseData),
-            checkRespondent2DQRequirements(caseData),
-            checkApplicant2DQRequirements(caseData),
-            checkLiPApplicantIsWelsh(caseData),
-            checkLiPDefendantIsWelsh(caseData)
-        ).anyMatch(Boolean::booleanValue);
+        return checkForActiveCaseFlags(caseData) || checkApplicant1DQRequirements(caseData)
+            || checkRespondent1DQRequirements(caseData) || checkRespondent2DQRequirements(caseData)
+            || checkApplicant2DQRequirements(caseData) || checkLiPApplicantIsWelsh(caseData)
+            || checkLiPDefendantIsWelsh(caseData);
     }
 
     private boolean checkLiPApplicantIsWelsh(CaseData caseData) {
-        log.info("Check LiP Application is Welsh for case ID {}", caseData.getCcdCaseReference());
         return caseData.isClaimantBilingual();
     }
 
     private boolean checkLiPDefendantIsWelsh(CaseData caseData) {
-        log.info("Check LiP Defendant is Welsh for case ID {}", caseData.getCcdCaseReference());
         return caseData.isRespondentResponseBilingual();
     }
 
     private boolean checkApplicant1DQRequirements(CaseData caseData) {
-        log.info("Check Applicant 1 DQ requirements for case ID {}", caseData.getCcdCaseReference());
-        if (caseData.getApplicant1DQ() == null) {
-            return false;
-        }
+        if (caseData.getApplicant1DQ() != null) {
+            Applicant1DQ applicant1DQ = caseData.getApplicant1DQ();
 
-        return Stream.of(
-            welshOrBilingualLanguage(caseData.getApplicant1DQ().getWelshLanguageRequirements()),
-            vulnerabilityAdjustmentRequired(caseData.getApplicant1DQ().getVulnerabilityQuestions()),
-            supportWithAccessNeedsRequired(caseData.getApplicant1DQ().getApplicant1DQHearingSupport())
-        ).anyMatch(Boolean::booleanValue);
+            return welshOrBilingualLanguage(applicant1DQ.getWelshLanguageRequirements())
+                || vulnerabilityAdjustmentRequired(applicant1DQ.getVulnerabilityQuestions())
+                || supportWithAccessNeedsRequired(applicant1DQ.getApplicant1DQHearingSupport());
+        }
+        return false;
     }
 
     private boolean checkApplicant2DQRequirements(CaseData caseData) {
-        log.info("Check Applicant 2 DQ requirements for case ID {}", caseData.getCcdCaseReference());
-        if (caseData.getApplicant2DQ() == null) {
-            return false;
-        }
+        if (caseData.getApplicant2DQ() != null) {
+            Applicant2DQ applicant2DQ = caseData.getApplicant2DQ();
 
-        return Stream.of(
-            welshOrBilingualLanguage(caseData.getApplicant2DQ().getWelshLanguageRequirements()),
-            vulnerabilityAdjustmentRequired(caseData.getApplicant2DQ().getVulnerabilityQuestions()),
-            supportWithAccessNeedsRequired(caseData.getApplicant2DQ().getApplicant2DQHearingSupport())
-        ).anyMatch(Boolean::booleanValue);
+            return welshOrBilingualLanguage(applicant2DQ.getWelshLanguageRequirements())
+                || vulnerabilityAdjustmentRequired(applicant2DQ.getVulnerabilityQuestions())
+                || supportWithAccessNeedsRequired(applicant2DQ.getApplicant2DQHearingSupport());
+        }
+        return false;
     }
 
     private boolean checkRespondent1DQRequirements(CaseData caseData) {
-        log.info("Check Respondent 1 DQ requirements for case ID {}", caseData.getCcdCaseReference());
-        if (caseData.getRespondent1DQ() == null) {
-            return false;
-        }
+        if (caseData.getRespondent1DQ() != null) {
+            Respondent1DQ respondent1DQ = caseData.getRespondent1DQ();
 
-        return Stream.of(
-            welshOrBilingualLanguage(caseData.getRespondent1DQ().getWelshLanguageRequirements()),
-            vulnerabilityAdjustmentRequired(caseData.getRespondent1DQ().getVulnerabilityQuestions()),
-            supportWithAccessNeedsRequired(caseData.getRespondent1DQ().getRespondent1DQHearingSupport())
-        ).anyMatch(Boolean::booleanValue);
+            return welshOrBilingualLanguage(respondent1DQ.getWelshLanguageRequirements())
+                || vulnerabilityAdjustmentRequired(respondent1DQ.getVulnerabilityQuestions())
+                || supportWithAccessNeedsRequired(respondent1DQ.getRespondent1DQHearingSupport());
+        }
+        return false;
     }
 
     private boolean checkRespondent2DQRequirements(CaseData caseData) {
-        log.info("Check Respondent 2 DQ requirements for case ID {}", caseData.getCcdCaseReference());
-        if (caseData.getRespondent2DQ() == null) {
-            return false;
-        }
+        if (caseData.getRespondent2DQ() != null) {
+            Respondent2DQ respondent2DQ = caseData.getRespondent2DQ();
 
-        return Stream.of(
-            welshOrBilingualLanguage(caseData.getRespondent2DQ().getWelshLanguageRequirements()),
-            vulnerabilityAdjustmentRequired(caseData.getRespondent2DQ().getVulnerabilityQuestions()),
-            supportWithAccessNeedsRequired(caseData.getRespondent2DQ().getRespondent2DQHearingSupport())
-        ).anyMatch(Boolean::booleanValue);
+            return welshOrBilingualLanguage(respondent2DQ.getWelshLanguageRequirements())
+                || vulnerabilityAdjustmentRequired(respondent2DQ.getVulnerabilityQuestions())
+                || supportWithAccessNeedsRequired(respondent2DQ.getRespondent2DQHearingSupport());
+        }
+        return false;
     }
 
     private boolean supportWithAccessNeedsRequired(HearingSupport hearingSupport) {
@@ -156,7 +154,6 @@ public class MediationJsonService {
     }
 
     private boolean checkForActiveCaseFlags(CaseData caseData) {
-        log.info("Check for active case flags for case ID {}", caseData.getCcdCaseReference());
         List<FlagDetail> allActiveCaseLevelFlags = getAllActiveCaseLevelFlags(caseData);
         List<PartyFlags> allActivePartyLevelFlags = getAllActiveFlags(caseData);
 
@@ -164,7 +161,6 @@ public class MediationJsonService {
     }
 
     private void buildLitigantFields(CaseData caseData, List<MediationLitigant> litigantList) {
-        log.info("Build litigant fields for case ID {}", caseData.getCcdCaseReference());
         litigantList.add(buildApplicant1Fields(caseData));
         if (caseData.getApplicant2() != null) {
             litigantList.add(buildApplicant2Fields(caseData));
@@ -176,7 +172,6 @@ public class MediationJsonService {
     }
 
     private MediationLitigant buildApplicant1Fields(CaseData caseData) {
-        log.info("Build Applicant 1 fields for case ID {}", caseData.getCcdCaseReference());
         if (NO.equals(caseData.getApplicant1Represented())) {
             return buildUnrepresentedLitigant(caseData.getApplicant1(),
                                               caseData.getCaseDataLiP().getApplicant1AdditionalLipPartyDetails() != null
@@ -193,7 +188,6 @@ public class MediationJsonService {
     }
 
     private MediationLitigant buildApplicant2Fields(CaseData caseData) {
-        log.info("Build Applicant 2 fields for case ID {}", caseData.getCcdCaseReference());
         return buildRepresentedLitigant(caseData.getApplicant2(),
                                         caseData.getApp1MediationContactInfo(), caseData.getApp1MediationAvailability(),
                                         caseData.getApplicant2OrganisationPolicy() != null
@@ -204,7 +198,6 @@ public class MediationJsonService {
     }
 
     private MediationLitigant buildRespondent1Fields(CaseData caseData) {
-        log.info("Build Respondent 1 fields for case ID {}", caseData.getCcdCaseReference());
         if (YES.equals(caseData.getRespondent1Represented())) {
             return buildRepresentedLitigant(caseData.getRespondent1(),
                                             caseData.getResp1MediationContactInfo(), caseData.getResp1MediationAvailability(),
@@ -219,7 +212,6 @@ public class MediationJsonService {
     }
 
     private MediationLitigant buildRespondent2Fields(CaseData caseData) {
-        log.info("Build Respondent 2 fields for case ID {}", caseData.getCcdCaseReference());
         if (YES.equals(caseData.getRespondent2SameLegalRepresentative())) {
             return buildRepresentedLitigant(caseData.getRespondent2(),
                                             caseData.getResp1MediationContactInfo(), caseData.getResp1MediationAvailability(),
@@ -234,10 +226,33 @@ public class MediationJsonService {
 
     private MediationLitigant buildUnrepresentedLitigant(Party party, String originalMediationContactPerson,
                                                          MediationLiPCarm mediationLiPCarm) {
-        var unrepresentedLitigantBuilder = MediationLitigant.builder();
-        partyDetailsPopulator.populator(unrepresentedLitigantBuilder, party);
-        unrepresentedLitigantPopulator.populator(unrepresentedLitigantBuilder, party, originalMediationContactPerson, mediationLiPCarm);
-        return unrepresentedLitigantBuilder.build();
+        List<MediationUnavailability> dateRangeToAvoid = getDateRangeToAvoid(mediationLiPCarm);
+
+        String mediationContactName = getUnrepresentedLitigantMediationContactName(
+            party, originalMediationContactPerson, mediationLiPCarm);
+
+        String mediationEmail = YES.equals(mediationLiPCarm.getIsMediationEmailCorrect())
+            ? party.getPartyEmail() : mediationLiPCarm.getAlternativeMediationEmail();
+
+        String mediationPhone = YES.equals(mediationLiPCarm.getIsMediationPhoneCorrect())
+            ? party.getPartyPhone() : mediationLiPCarm.getAlternativeMediationTelephone();
+
+        String partyRole = party.getFlags() != null ? party.getFlags().getRoleOnCase() : null;
+        return MediationLitigant.builder()
+            .partyID(party.getPartyID())
+            .partyRole(partyRole)
+            .partyType(party.getType())
+            .partyName(party.getPartyName())
+            .paperResponse(PAPER_RESPONSE)
+            .represented(false)
+            .solicitorOrgName(null)
+            .litigantEmail(party.getPartyEmail())
+            .litigantTelephone(party.getPartyPhone())
+            .mediationContactName(mediationContactName)
+            .mediationContactNumber(mediationPhone)
+            .mediationContactEmail(mediationEmail)
+            .dateRangeToAvoid(dateRangeToAvoid)
+            .build();
     }
 
     private MediationLitigant buildRepresentedLitigant(Party party,
@@ -245,10 +260,99 @@ public class MediationJsonService {
                                                        MediationAvailability mediationAvailability,
                                                        OrganisationPolicy organisationPolicy, String solicitorEmail) {
 
-        var representedLitigantBuilder = MediationLitigant.builder();
-        partyDetailsPopulator.populator(representedLitigantBuilder, party);
-        representedLitigantPopulator.populator(representedLitigantBuilder, mediationContactInformation, mediationAvailability, organisationPolicy, solicitorEmail);
-        return representedLitigantBuilder.build();
+        String solicitorOrgName = getSolicitorOrgName(organisationPolicy);
+        String partyRole = party.getFlags() != null ? party.getFlags().getRoleOnCase() : null;
+        String mediationContactName = getMediationContactName(mediationContactInformation);
+        String mediationContactNumber = mediationContactInformation != null
+            ? mediationContactInformation.getTelephoneNumber() : null;
+        String mediationEmail = mediationContactInformation != null
+            ? mediationContactInformation.getEmailAddress() : null;
+
+        List<MediationUnavailability> dateRangeToAvoid = getDateRangeToAvoid(mediationAvailability);
+
+        return MediationLitigant.builder()
+            .partyID(party.getPartyID())
+            .partyRole(partyRole)
+            .partyType(party.getType())
+            .partyName(party.getPartyName())
+            .paperResponse(PAPER_RESPONSE)
+            .represented(true)
+            .solicitorOrgName(solicitorOrgName)
+            .litigantEmail(solicitorEmail)
+            .litigantTelephone(null)
+            .mediationContactName(mediationContactName)
+            .mediationContactNumber(mediationContactNumber)
+            .mediationContactEmail(mediationEmail)
+            .dateRangeToAvoid(dateRangeToAvoid)
+            .build();
     }
 
+    private List<MediationUnavailability> toMediationUnavailableDates(List<Element<UnavailableDate>> unavailableDatesForMediation) {
+        List<UnavailableDate> unavailableDates = unwrapElements(unavailableDatesForMediation);
+        List<MediationUnavailability> toMediationUnavailability = new ArrayList<>();
+        for (UnavailableDate unavailableDate : unavailableDates) {
+            if (SINGLE_DATE.equals(unavailableDate.getUnavailableDateType())) {
+                toMediationUnavailability.add(MediationUnavailability.builder()
+                                                  .dateFrom(formatDate(unavailableDate.getDate()))
+                                                  .dateTo(formatDate(unavailableDate.getDate()))
+                                                  .build());
+            }
+            if (DATE_RANGE.equals(unavailableDate.getUnavailableDateType())) {
+                toMediationUnavailability.add(MediationUnavailability.builder()
+                                                  .dateFrom(formatDate(unavailableDate.getFromDate()))
+                                                  .dateTo(formatDate(unavailableDate.getToDate()))
+                                                  .build());
+            }
+        }
+        return toMediationUnavailability;
+    }
+
+    private List<MediationUnavailability> getDateRangeToAvoid(MediationAvailability mediationAvailability) {
+        if (mediationAvailability != null && YES.equals(mediationAvailability.getIsMediationUnavailablityExists())) {
+            return toMediationUnavailableDates(mediationAvailability.getUnavailableDatesForMediation());
+        }
+        return List.of(MediationUnavailability.builder().build());
+    }
+
+    private List<MediationUnavailability> getDateRangeToAvoid(MediationLiPCarm mediationLiPCarm) {
+        if (YES.equals(mediationLiPCarm.getHasUnavailabilityNextThreeMonths())) {
+            return toMediationUnavailableDates(mediationLiPCarm.getUnavailableDatesForMediation());
+        }
+        return List.of(MediationUnavailability.builder().build());
+    }
+
+    private String formatDate(LocalDate unavailableDate) {
+        return unavailableDate.format(DateTimeFormatter.ofPattern(DATE_FORMAT, Locale.UK));
+    }
+
+    @Nullable
+    private String getMediationContactName(MediationContactInformation mediationContactInformation) {
+        return mediationContactInformation != null
+            ? String.format("%s %s", mediationContactInformation.getFirstName(), mediationContactInformation.getLastName())
+            : null;
+    }
+
+    @Nullable
+    private String getSolicitorOrgName(OrganisationPolicy organisationPolicy) {
+        String orgId = organisationPolicy.getOrganisation().getOrganisationID();
+        Optional<Organisation> organisation = organisationService.findOrganisationById(orgId);
+        Organisation solicitorOrgDetails = organisation.orElse(null);
+        return solicitorOrgDetails != null ? solicitorOrgDetails.getName() : null;
+    }
+
+    private boolean isIndividualOrSoleTrader(Party party) {
+        return Party.Type.INDIVIDUAL.equals(party.getType())
+            || Party.Type.SOLE_TRADER.equals(party.getType());
+    }
+
+    private String getUnrepresentedLitigantMediationContactName(Party party,
+                                                                 String originalMediationContactPerson,
+                                                                 MediationLiPCarm mediationLiPCarm) {
+        if (isIndividualOrSoleTrader(party)) {
+            return party.getPartyName();
+        } else {
+            return YES.equals(mediationLiPCarm.getIsMediationContactNameCorrect())
+                ? originalMediationContactPerson : mediationLiPCarm.getAlternativeMediationContactPerson();
+        }
+    }
 }
